Moved gameVisiblityList to it's own folder
